+++ src/Writer.php
@@ -13,9 +13,6 @@
 namespace League\Csv;
 
 use InvalidArgumentException;
-use OutOfBoundsException;
-use RuntimeException;
-use SplFileObject;
 use Traversable;
 
 /**
@@ -27,47 +24,6 @@ use Traversable;
  */
 class Writer extends AbstractCsv
 {
-    /**
-     * set null handling mode to throw exception
-     */
-    const NULL_AS_EXCEPTION = 1;
-
-    /**
-     * set null handling mode to remove cell
-     */
-    const NULL_AS_SKIP_CELL = 2;
-
-    /**
-     * set null handling mode to convert null into empty string
-     */
-    const NULL_AS_EMPTY = 3;
-
-    /**
-     * disable null handling
-     */
-    const NULL_HANDLING_DISABLED = 4;
-
-    /**
-     * the object current null handling mode
-     *
-     * @var int
-     */
-    protected $null_handling_mode = self::NULL_AS_EXCEPTION;
-
-    /**
-     * The number of column per row
-     *
-     * @var int
-     */
-    protected $columns_count = -1;
-
-    /**
-     * should the class detect the column count based the inserted row
-     *
-     * @var bool
-     */
-    protected $detect_columns_count = false;
-
     /**
      * {@ihneritdoc}
      */
@@ -81,89 +37,19 @@ class Writer extends AbstractCsv
     protected $csv;
 
     /**
+     * Callables to filter the iterator
+     *
+     * @var callable[]
+     */
+    protected $rules = [];
+
+    /**
      * should the class validate the input before insertion
      *
      * @var boolean
      */
     protected $useValidation = true;
 
-    /**
-     * Tell the class how to handle null value
-     *
-     * @param int $value a Writer null behavior constant
-     *
-     * @throws \OutOfBoundsException If the Integer is not valid
-     *
-     * @return static
-     */
-    public function setNullHandlingMode($value)
-    {
-        if (! in_array($value, [
-                self::NULL_AS_SKIP_CELL,
-                self::NULL_AS_EXCEPTION,
-                self::NULL_AS_EMPTY,
-                self::NULL_HANDLING_DISABLED,
-            ])) {
-            throw new OutOfBoundsException('invalid value for null handling');
-        }
-        $this->null_handling_mode = $value;
-
-        return $this;
-    }
-
-    /**
-     * null handling getter
-     *
-     * @return int
-     */
-    public function getNullHandlingMode()
-    {
-        return $this->null_handling_mode;
-    }
-
-    /**
-     * Set Inserted row column count
-     *
-     * @param int $value
-     *
-     * @throws \InvalidArgumentException If $value is lesser than -1
-     *
-     * @return static
-     */
-    public function setColumnsCount($value)
-    {
-        if (false === filter_var($value, FILTER_VALIDATE_INT, ['options' => ['min_range' => -1]])) {
-            throw new InvalidArgumentException('the column count must an integer greater or equals to -1');
-        }
-        $this->detect_columns_count = false;
-        $this->columns_count = $value;
-
-        return $this;
-    }
-
-    /**
-     * Column count getter
-     *
-     * @return int
-     */
-    public function getColumnsCount()
-    {
-        return $this->columns_count;
-    }
-
-    /**
-     * The method will set the $columns_count property according to the next inserted row
-     * and therefore will also validate the next line whatever length it has no matter
-     * the current $columns_count property value.
-     *
-     * @return static
-     */
-    public function autodetectColumnsCount()
-    {
-        $this->detect_columns_count = true;
-
-        return $this;
-    }
 
     /**
      * Tells wether the library should check or not the input
@@ -172,11 +58,86 @@ class Writer extends AbstractCsv
      *
      * @return static
      */
-    public function useValidation($status)
+    public function useValidation($activate)
     {
-        $this->useValidation = (bool) $status;
+        $this->useValidation = (bool) $activate;
 
         return $this;
+    }
+
+    /**
+     * Set an Iterator sorting callable function
+     *
+     * @param callable $callable
+     *
+     * @return static
+     */
+    public function addValidationRule(callable $callable)
+    {
+        $this->rules[] = $callable;
+
+        return $this;
+    }
+
+    /**
+     * Remove a callable from the collection
+     *
+     * @param callable $callable
+     *
+     * @return static
+     */
+    public function removeValidationRule(callable $callable)
+    {
+        $res = array_search($callable, $this->rules, true);
+        if (false !== $res) {
+            unset($this->rules[$res]);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Detect if the callable is already registered
+     *
+     * @param callable $callable
+     *
+     * @return bool
+     */
+    public function hasValidationRule(callable $callable)
+    {
+        return false !== array_search($callable, $this->rules, true);
+    }
+
+    /**
+     * Remove all registered callable
+     *
+     * @return static
+     */
+    public function clearValidationRules()
+    {
+        $this->rules = [];
+
+        return $this;
+    }
+
+    /**
+    * Filter the Iterator
+    *
+    * @param array $row
+    *
+    * @return array
+    */
+    protected function applyValidationRules(array $row)
+    {
+        if (! $this->useValidation || ! $this->rules) {
+            return $row;
+        }
+
+        foreach ($this->rules as $rule) {
+            $row = $rule($row);
+        }
+
+        return $row;
     }
 
     /**
@@ -215,9 +176,7 @@ class Writer extends AbstractCsv
     public function insertOne($row)
     {
         $row = $this->formatRow($row);
-        if ($this->useValidation) {
-            $row = $this->validateRow($row);
-        }
+        $row = $this->applyValidationRules($row);
         $csv = $this->getCsv();
         $csv->fputcsv($row, $this->delimiter, $this->enclosure);
         if ("\n" !== $this->newline) {
@@ -242,91 +201,6 @@ class Writer extends AbstractCsv
         }
 
         return $row;
-    }
-
-    /**
-     * Is the submitted row valid
-     *
-     * @param array $row
-     *
-     * @throws \InvalidArgumentException If the given $row is not valid
-     * @throws \RuntimeException If the given $row does not contain valid column count
-     *
-     * @return array
-     */
-    protected function validateRow(array $row)
-    {
-        if (self::NULL_HANDLING_DISABLED != $this->null_handling_mode) {
-            array_walk($row, function ($value) {
-                if (! $this->isConvertibleContent($value)) {
-                    throw new InvalidArgumentException('The values are not convertible into strings');
-                }
-            });
-            $row = $this->sanitizeColumnsContent($row);
-        }
-
-        if (! $this->isColumnsCountConsistent($row)) {
-            throw new RuntimeException('Adding '.count($row).' cells on a {$this->columns_count} cells per row CSV.');
-        }
-
-        return $row;
-    }
-
-    /**
-     * Check if a given value can be added into a CSV cell
-     *
-     * The value MUST respect the null handling mode
-     * The value MUST be convertible into a string
-     *
-     * @param string|null $value the value to be added
-     *
-     * @return bool
-     */
-    protected function isConvertibleContent($value)
-    {
-        return (is_null($value) && self::NULL_AS_EXCEPTION != $this->null_handling_mode)
-            || self::isValidString($value);
-    }
-
-    /**
-     * Format the row according to the null handling behavior
-     *
-     * @param array $row
-     *
-     * @return array
-     */
-    protected function sanitizeColumnsContent(array $row)
-    {
-        if (self::NULL_AS_EMPTY == $this->null_handling_mode) {
-            return str_replace(null, '', $row);
-        } elseif (self::NULL_AS_SKIP_CELL == $this->null_handling_mode) {
-            return array_filter($row, function ($value) {
-                return !is_null($value);
-            });
-        }
-
-        return $row;
-    }
-
-    /**
-     * Check column count consistency
-     *
-     * @param array $row the row to be added to the CSV
-     *
-     * @return bool
-     */
-    protected function isColumnsCountConsistent(array $row)
-    {
-        if ($this->detect_columns_count) {
-            $this->columns_count = count($row);
-            $this->detect_columns_count = false;
-
-            return true;
-        } elseif (-1 == $this->columns_count) {
-            return true;
-        }
-
-        return count($row) == $this->columns_count;
     }
 
     /**